Add script for reconstructing tmin and tmax
Eventually we want to reconstruct/predict daily tmin and tmax across
landscapes. But we have thousands of days of historical data (and
predicted PCA scores), and thousands of geographic coordinates (and
predicted PCA loadings), so the resulting matrix is too big for R.

I need to open an issue as I am not sure about the solution./ Perhaps we
cannot map out tmin and tmax across the landscape like we wanted.