Update after new clipped topo grids added.
Fixes #37

We now are using smaller topo grids clipped to a more manageable area of
interest for each mtn range.

The data is currently on dropbox but we could consider adding to the repo.

Here are @hpoulos's notes on the clipping method using arcgis:

Clip NLCD (National Land Cover Dataset) canopy cover raster to
the extents of the HUC-clipped rasters for each mountain range to
minimize NLCD file size. Reclassify raster to No Data or class 1
= canopy cover > 10%. In conversion tools, use the raster to
polygon tool. Be sure to that the simplify polygons box remains
unchecked or it will not work. Then use the minimum bounding
geometry operation. Select geometry type as “Convex Hulls” and
hit OK. Select the large polygon if multiple smaller polygons are
also generated (all of these small polygons were outside the
extent of the original HUC-clipped grids anyway), and export as a
new shapefile. Then, in Model Builder, clip each old .asc raster
dataset to the extent of the convex hull shapefile, and export as
a new .asc file.